Feat : 게시판 상세 조회 API 일부 구현

+++ src/main/java/com/acorn/api/service/board/BoardService.java
@@ -16,5 +16,5 @@ public interface BoardService {
 
     void boardDataSave(BoardSaveDTO boardSaveDTO);
 
-    BoardDetailDTO getBoardDetailData(Long boardId);
+    BoardDetailDTO getBoardDetailData(Integer boardId);
 }
